# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SoccerWatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recordings/{id}:
    get:
      operationId: getRecording
      tags: [recordings]
      summary: Get a recording by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "404":
          description: Not found
...
  /recordings/{id}/claim-match/off-pitch:
    post:
      operationId: createClaimMatchOffPitchSpan
      tags: [claimMatch]
      summary: Declare a tracking-time period with no playable pitch action
      security:
        - clerkSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimOffPitchInput"
      responses:
        "201":
          description: Off-pitch period recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimOffPitchSpan"
        "200":
          description: Idempotent replay of an existing period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimOffPitchSpan"
        "400":
          description: Invalid tracking-time period
        "401":
          description: Authenticated account required
        "404":
          description: Recording or tracking bundle not found
        "409":
          description: The period overlaps vouched identity fragments and requires explicit confirmation
  /recordings/{id}/claim-match/off-pitch/{clientId}:
    delete:
      operationId: deleteClaimMatchOffPitchSpan
      tags: [claimMatch]
      summary: Delete an off-pitch period by client id
      security:
        - clerkSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Off-pitch period deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimOffPitchDeleteResponse"
        "400":
          description: Invalid recording or client id
        "401":
          description: Authenticated account required
        "404":
          description: Off-pitch period not found
...
components:
  schemas:
...
    Recording:
      type: object
      required: [id, fieldId, court, date, timeSlot, duration]
      properties:
        id:
          type: integer
        fieldId:
          type: integer
        court:
          type: string
        date:
          type: string
        timeSlot:
          type: string
        duration:
          type: string
        score:
          type: ["string", "null"]
        videoUrl:
          type: string
        highlightMoment:
          type: ["string", "null"]
        fieldName:
          type: ["string", "null"]
...
    ClaimOffPitchInput:
      type: object
      required: [clientId, fromSeconds, toSeconds]
      properties:
        clientId:
          type: string
          minLength: 1
        fromSeconds:
          type: number
          minimum: 0
          description: Exact tracking-time seconds from the start of the tracking window.
        toSeconds:
          type: number
          minimum: 0
          description: Exact tracking-time seconds from the start of the tracking window.
        confirmConflict:
          type: boolean
          description: Explicitly acknowledge overlap with vouched identity fragments.
    ClaimOffPitchSpan:
      type: object
      required: [id, clientId, fromSeconds, toSeconds, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: string
        fromSeconds:
          type: number
        toSeconds:
          type: number
        createdAt:
          type: string
    ClaimOffPitchDeleteResponse:
      type: object
      required: [deleted, clientId]
      properties:
        deleted:
          type: boolean
        clientId:
          type: string
...
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      bearerFormat: JWT